added classess for VNP14A1
